@@ -1,6 +1,6 @@
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
-from models import db
+from Models import db
 from run import create_app
 
 app = create_app('config')
