@@ -1,8 +1,8 @@
 """empty message
 
-Revision ID: 293ce22fe58d
+Revision ID: 2568c1256eff
 Revises: 
-Create Date: 2020-06-25 16:03:55.585194
+Create Date: 2021-02-05 16:07:11.455372
 
 """
 from alembic import op
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 
 
 # revision identifiers, used by Alembic.
-revision = '293ce22fe58d'
+revision = '2568c1256eff'
 down_revision = None
 branch_labels = None
 depends_on = None
@@ -18,14 +18,31 @@ depends_on = None
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('groups',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=True),
+    sa.Column('group_key', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
     op.create_table('users',
     sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('username', sa.String(), nullable=True),
-    sa.Column('password', sa.String(), nullable=True),
-    sa.Column('emailaddress', sa.String(), nullable=True),
+    sa.Column('firstname', sa.String(), nullable=True),
+    sa.Column('lastname', sa.String(), nullable=True),
+    sa.Column('phonenumber', sa.String(length=15), nullable=True),
+    sa.Column('username', sa.String(length=64), nullable=True),
+    sa.Column('password', sa.String(length=128), nullable=True),
+    sa.Column('emailaddress', sa.String(length=120), nullable=True),
     sa.Column('api_key', sa.String(), nullable=True),
+    sa.Column('avatar', sa.LargeBinary(), nullable=True),
     sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('phonenumber'),
     sa.UniqueConstraint('username')
+    )
+    op.create_table('group_member',
+    sa.Column('group_id', sa.Integer(), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['group_id'], ['groups.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], )
     )
     op.create_table('tasks',
     sa.Column('id', sa.Integer(), nullable=False),
@@ -38,7 +55,7 @@ def upgrade():
     sa.Column('reminders', sa.String(), nullable=True),
     sa.Column('group', sa.String(), nullable=True),
     sa.Column('task_key', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
     sa.PrimaryKeyConstraint('id')
     )
     op.create_table('subtasks',
@@ -52,7 +69,7 @@ def upgrade():
     sa.Column('group', sa.String(), nullable=True),
     sa.Column('index', sa.Integer(), nullable=True),
     sa.Column('subtask_key', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ),
+    sa.ForeignKeyConstraint(['task_id'], ['tasks.id'], ondelete='CASCADE'),
     sa.PrimaryKeyConstraint('id')
     )
     # ### end Alembic commands ###
@@ -62,5 +79,7 @@ def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.drop_table('subtasks')
     op.drop_table('tasks')
+    op.drop_table('group_member')
     op.drop_table('users')
+    op.drop_table('groups')
     # ### end Alembic commands ###
