@@ -8,7 +8,7 @@ def create_app(config_filename):
     from app import api_bp
     app.register_blueprint(api_bp, url_prefix='/api')
 
-    from models import db
+    from Models import db
     db.init_app(app)
 
     return app
