@@ -5,22 +5,56 @@ import 'package:todolist/models/global.dart';
 
 class TitleCard extends StatelessWidget {
   final String title;
-  final VoidCallback addTask;
+  final Widget child;
+  //final VoidCallback addTask;
 
-  TitleCard(this.title, this.addTask);
+  TitleCard({this.title, this.child});
   @override
   Widget build(BuildContext context) {
     return Stack(
       children: <Widget>[
-        _TitleCardBackground(),
-        _TitleCardTitle(title),
-        _TitleCardButton(addTask),
+        child,
+        _titleCardBackground(),
+        //_titleCardTitle(),
+        //_TitleCardButton(addTask),
       ],
+    );
+  }
+
+  Container _titleCardBackground() {
+    return Container(
+      height: 120,
+      width: double.infinity,
+      decoration: new BoxDecoration(
+        color: Colors.white,
+        borderRadius: BorderRadius.only(
+          bottomLeft: Radius.circular(40),
+          bottomRight: Radius.circular(40),
+        ),
+        boxShadow: [
+          new BoxShadow(
+            color: Colors.black.withOpacity(0.7),
+            blurRadius: 25.0,
+          ),
+        ],
+      ),
+      child: _titleCardTitle(),
+      padding: EdgeInsets.only(bottom: 20, left: 15.0),
+    );
+  }
+
+  Align _titleCardTitle() {
+    return Align(
+      alignment: Alignment.bottomLeft,
+      child: Text(
+        title,
+        style: cardTitleStyle,
+      ),
     );
   }
 }
 
-class _TitleCardTitle extends StatelessWidget {
+/* class _TitleCardTitle extends StatelessWidget {
   final String title;
   _TitleCardTitle(this.title);
   @override
@@ -35,9 +69,9 @@ class _TitleCardTitle extends StatelessWidget {
       ),
     );
   }
-}
+} */
 
-class _TitleCardBackground extends StatelessWidget {
+/* class _TitleCardBackground extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -51,9 +85,9 @@ class _TitleCardBackground extends StatelessWidget {
       ),
     );
   }
-}
+} */
 
-class _TitleCardButton extends StatefulWidget {
+/* class _TitleCardButton extends StatefulWidget {
   final VoidCallback addTaskDialog;
   
 
@@ -77,9 +111,10 @@ class __TitleCardButtonState extends State<_TitleCardButton> {
 
   @override
   Widget build(BuildContext context) {
-    return Transform.scale(
-      scale: 1.1,
-      origin: Offset(-1800, -950),
+    return Align(
+      /* scale: 1.1,
+      origin: Offset(-3300, -500) */
+      alignment: Alignment(0.9, 0.95),
       child: FloatingActionButton(
         tooltip: "Press to Add a Task",
         onPressed: () {
@@ -89,4 +124,4 @@ class __TitleCardButtonState extends State<_TitleCardButton> {
       ),
     );
   }
-}
+} */
