@@ -2,6 +2,8 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' show Client;
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:todolist/models/group.dart';
+import 'package:todolist/models/groupmember.dart';
 import 'package:todolist/models/subtasks.dart';
 import 'package:todolist/models/tasks.dart';
 import 'dart:convert';
@@ -9,8 +11,18 @@ import 'package:todolist/models/user.dart';
 
 class ApiProvider {
   Client client = Client();
-  // final _apiKey = 'your_api_key';
+  static String baseURL = "http://10.0.2.2:5000/api";
+  String signinURL = baseURL + "/signin";
+  String userURL = baseURL + "/user";
+  String taskURL = baseURL + "/tasks";
+  String subtaskURL = baseURL + "/subtasks";
+  String groupURL = baseURL + "/group";
+  String groupmemberURL = baseURL + "/groupmember";
 
+  String apiKey;
+
+//User CRUD Functions
+  //Sign Up
   Future<User> registerUser(
       String username,
       String password,
@@ -19,7 +31,7 @@ class ApiProvider {
       String lastname,
       String phonenumber,
       ImageProvider avatar) async {
-    final response = await client.post("http://10.0.2.2:5000/api/user",
+    final response = await client.post(userURL,
         // headers: "",
         body: jsonEncode({
           "emailaddress": email,
@@ -41,69 +53,26 @@ class ApiProvider {
     }
   }
 
+  //Sign In
   Future signinUser(String username, String password, String apiKey) async {
-    final response = await client.post("http://10.0.2.2:5000/api/signin",
+    final response = await client.post(signinURL,
         headers: {"Authorization": apiKey},
         body: jsonEncode({
           "username": username,
           "password": password,
         }));
     final Map result = json.decode(response.body);
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
       await saveApiKey(result["data"]["api_key"]);
       return User.fromJson(result["data"]);
     } else {
       // If that call was not successful, throw an error.
-      throw Exception('Failed to load post');
+      throw Exception(result["message"]);
     }
   }
 
-  Future<List<Task>> getUserTasks(String apiKey) async {
-    final response = await client.get(
-      "http://10.0.2.2:5000/api/tasks",
-      headers: {"Authorization": apiKey},
-    );
-    final Map result = json.decode(response.body);
-    if (response.statusCode == 201) {
-      // If the call to the server was successful, parse the JSON
-      List<Task> tasks = [];
-      for (Map json_ in result["data"]) {
-        try {
-          tasks.add(Task.fromJson(json_));
-        } catch (Exception) {
-          print(Exception);
-        }
-      }
-      return tasks;
-    } else {
-      // If that call was not successful, throw an error.
-      throw Exception('Failed to load tasks');
-    }
-  }
-
-  Future addUserTask(
-      String apiKey, String taskName, String groupName, int index) async {
-    final response = await client.post("http://10.0.2.2:5000/api/tasks",
-        headers: {"Authorization": apiKey},
-        body: jsonEncode({
-          "title": taskName,
-          "note": "",
-          "completed": false,
-          "repeats": "",
-          "group": groupName,
-          "reminders": "",
-          "index": index
-        }));
-    if (response.statusCode == 201) {
-      print("Task " + taskName + " added");
-    } else {
-      // If that call was not successful, throw an error.
-      print(json.decode(response.body));
-      throw Exception('Failed to post tasks');
-    }
-  }
-
+  // Edit Profile
   Future updateUserProfile(
       String currentPassword,
       String newPassword,
@@ -112,9 +81,8 @@ class ApiProvider {
       String firstname,
       String lastname,
       String phonenumber,
-      ImageProvider avatar,
-      String apiKey) async {
-    final response = await client.put("http://10.0.2.2:5000/api/user",
+      ImageProvider avatar) async {
+    final response = await client.put(userURL,
         headers: {"Authorization": apiKey},
         body: jsonEncode({
           "currentPassword": currentPassword,
@@ -127,7 +95,7 @@ class ApiProvider {
           "avatar": avatar,
         }));
     final Map result = json.decode(response.body);
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       print("User Profile Updated");
       return User.fromJson(result["data"]);
     } else {
@@ -137,19 +105,122 @@ class ApiProvider {
     }
   }
 
-  Future updateUserTask(Task task) async {
-    final response = await client.put("http://10.0.2.2:5000/api/tasks",
+//Group CRUD Functions
+  // Get User's Group's
+  Future<List<Group>> getUserGroups() async {
+    final _apiKey = await getApiKey();
+    final response = await client.get(
+      groupURL,
+      headers: {"Authorization": _apiKey},
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 200) {
+      // If the call to the server was successful, parse the JSON
+      List<Group> groups = [];
+      for (Map json_ in result["data"]) {
+        try {
+          Group group = Group.fromJson(json_);
+          group.members = await getGroupMembers(group.groupKey);
+          group.tasks = await getTasks(group.groupKey);
+          groups.add(group);
+        } catch (Exception) {
+          print(Exception);
+        }
+      }
+      return groups;
+    } else {
+      // If that call was not successful, throw an error.
+      throw Exception(result["Message"]);
+    }
+  }
+
+  //GroupMember CRUD Functions
+  // Get Group's Members
+  Future<List<GroupMember>> getGroupMembers(String groupKey) async {
+    final response = await client.get(
+      groupmemberURL,
+      headers: {"Authorization": groupKey},
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 200) {
+      // If the call to the server was successful, parse the JSON
+      List<GroupMember> groupMembers = [];
+      for (Map json_ in result["data"]) {
+        try {
+          groupMembers.add(GroupMember.fromJson(json_));
+        } catch (Exception) {
+          print(Exception);
+          throw Exception;
+        }
+      }
+      return groupMembers;
+    } else {
+      // If that call was not successful, throw an error.
+      throw Exception(result["Message"]);
+    }
+  }
+
+//Task CRUD Functions
+  // Get Tasks
+  Future<List<Task>> getTasks(String groupKey) async {
+    final response = await client.get(
+      taskURL,
+      headers: {"Authorization": groupKey},
+    );
+    final Map result = json.decode(response.body);
+    if (response.statusCode == 200) {
+      // If the call to the server was successful, parse the JSON
+      List<Task> tasks = [];
+      for (Map json_ in result["data"]) {
+        try {
+          Task task = Task.fromJson(json_);
+          task.subtasks = await getSubtasks(task.taskKey);
+          tasks.add(task);
+        } catch (Exception) {
+          print(Exception);
+        }
+      }
+      return tasks;
+    } else {
+      // If that call was not successful, throw an error.
+      throw Exception(result["Message"]);
+    }
+  }
+
+  //Add Task
+  Future addTask(
+      String taskName, String groupKey, int index, bool completed) async {
+    final response = await client.post(taskURL,
+        headers: {"Authorization": apiKey},
+        body: jsonEncode({
+          "title": taskName,
+          "group_key": groupKey,
+          "index": index,
+          "completed": completed
+        }));
+    if (response.statusCode == 201) {
+      print("Task " + taskName + " added");
+    } else {
+      // If that call was not successful, throw an error.
+      final Map result = json.decode(response.body);
+      print(result["Message"]);
+      throw Exception(result["Message"]);
+    }
+  }
+
+  //Update Task
+  Future updateTask(Task task) async {
+    final response = await client.put(taskURL,
         headers: {"Authorization": task.taskKey},
         body: jsonEncode({
           "title": task.title,
           "note": task.note,
           "repeats": task.repeats,
           "completed": task.completed,
-          "group": task.group,
           "reminders": task.reminders,
           "index": task.index
         }));
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       print("Task " + task.title + " Updated");
     } else {
       // If that call was not successful, throw an error.
@@ -158,33 +229,36 @@ class ApiProvider {
     }
   }
 
-  Future deleteUserTask(String taskKey) async {
+  //Delete Task
+  Future deleteTask(String taskKey) async {
     final response = await client.delete(
-      "http://10.0.2.2:5000/api/tasks",
+      taskURL,
       headers: {"Authorization": taskKey},
     );
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       // If the call to the server was successful
       print("Task deleted");
     } else {
       // If that call was not successful, throw an error.
-      print(json.decode(response.body));
-      throw Exception('Failed to delete task');
+      final Map result = json.decode(response.body);
+      throw Exception(result["Message"]);
     }
   }
 
-  Future<List<SubTask>> getSubTasks(String taskKey) async {
+//Subtask CRUD Functions
+  //Get Subtasks
+  Future<List<Subtask>> getSubtasks(String taskKey) async {
     final response = await client.get(
-      "http://10.0.2.2:5000/api/subtasks",
+      subtaskURL,
       headers: {"Authorization": taskKey},
     );
     final Map result = json.decode(response.body);
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       // If the call to the server was successful, parse the JSON
-      List<SubTask> subtasks = [];
+      List<Subtask> subtasks = [];
       for (Map json_ in result["data"]) {
         try {
-          subtasks.add(SubTask.fromJson(json_));
+          subtasks.add(Subtask.fromJson(json_));
         } catch (Exception) {
           print(Exception);
         }
@@ -192,17 +266,18 @@ class ApiProvider {
       return subtasks;
     } else {
       // If that call was not successful, throw an error.
-      throw Exception('Failed to load tasks');
+      throw Exception(result["Message"]);
     }
   }
 
-  Future addSubTask(String taskKey, String subtaskName, String notes, int index,
-      bool completed) async {
-    final response = await client.post("http://10.0.2.2:5000/api/subtasks",
+  //Add Subtask
+  Future addSubtask(
+      String taskKey, String subtaskName, int index, bool completed) async {
+    final response = await client.post(subtaskURL,
         headers: {"Authorization": taskKey},
         body: jsonEncode({
           "title": subtaskName,
-          "note": notes,
+          "note": "",
           "completed": completed,
           "repeats": "",
           "group": "",
@@ -210,16 +285,17 @@ class ApiProvider {
           "index": index
         }));
     if (response.statusCode == 201) {
-      print("SubTask " + subtaskName + " added");
+      print("Subtask " + subtaskName + " added");
     } else {
       // If that call was not successful, throw an error.
-      print(json.decode(response.body));
-      throw Exception('Failed to load tasks');
+      final Map result = json.decode(response.body);
+      throw Exception(result["Message"]);
     }
   }
 
-  Future updateSubTask(SubTask subtask) async {
-    final response = await client.put("http://10.0.2.2:5000/api/subtasks",
+  //Update Subtask
+  Future updateSubtask(Subtask subtask) async {
+    final response = await client.put(subtaskURL,
         headers: {"Authorization": subtask.subtaskKey},
         body: jsonEncode({
           "title": subtask.title,
@@ -230,32 +306,41 @@ class ApiProvider {
           "reminders": subtask.reminders,
           "index": subtask.index
         }));
-    if (response.statusCode == 201) {
-      print("SubTask " + subtask.title + " Updated");
+    if (response.statusCode == 200) {
+      print("Subtask " + subtask.title + " Updated");
     } else {
       // If that call was not successful, throw an error.
-      print(json.decode(response.body));
-      throw Exception('Failed to update tasks');
+      final Map result = json.decode(response.body);
+      throw Exception(result["Message"]);
     }
   }
 
-  Future deleteSubTask(String subtaskKey) async {
+  //Delete Subtask
+  Future deleteSubtask(String subtaskKey) async {
     final response = await client.delete(
-      "http://10.0.2.2:5000/api/subtasks",
+      subtaskURL,
       headers: {"Authorization": subtaskKey},
     );
-    if (response.statusCode == 201) {
+    if (response.statusCode == 200) {
       // If the call to the server was successful
-      print("SubTask deleted");
+      print("Subtask deleted");
     } else {
       // If that call was not successful, throw an error.
-      print(json.decode(response.body));
-      throw Exception('Failed to delete subtask');
+      final Map result = json.decode(response.body);
+      throw Exception(result["Message"]);
     }
   }
 
+  //Save API key
   saveApiKey(String apiKey) async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     await prefs.setString('API_Token', apiKey);
+    this.apiKey = apiKey;
+  }
+
+  Future<String> getApiKey() async {
+    //if(apiKey.isNotEmpty) return apiKey;
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    return prefs.getString('API_Token');
   }
 }
