@@ -9,6 +9,13 @@ Color red = new Color(0xfff20f38);
 Color darkRed = new Color(0xffbf0426);
 Color brownRed = new Color(0xff590209);
 
+final appTitleStyle = TextStyle(
+  color: darkBlueGradient,
+  fontWeight: FontWeight.bold,
+  fontFamily: "Segoe UI",
+  fontSize: 28.0,
+);
+
 final hintTextStyle = TextStyle(
   color: Colors.white54,
   fontFamily: 'OpenSans',
@@ -45,50 +52,42 @@ final profileBoxDecorationStyle = BoxDecoration(
 );
 
 TextStyle cardTitleStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.bold, 
-  color: darkRed, 
-  fontSize: 50
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.bold,
+    color: darkRed,
+    fontSize: 50);
 
 TextStyle toDoListTileStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.w500, 
-  color: Colors.white, 
-  fontSize: 30
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.w500,
+    color: Colors.white,
+    fontSize: 30);
 
 TextStyle toDoListSubtitleStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.w300, 
-  color: Colors.white, 
-  fontSize: 18
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.w300,
+    color: Colors.white,
+    fontSize: 18);
 
 TextStyle taskListTitleStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.bold, 
-  color: darkBlue, 
-  fontSize: 50
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.bold,
+    color: darkBlue,
+    fontSize: 50);
 
 TextStyle loginTitleStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.bold, 
-  color: darkBlueGradient, 
-  fontSize: 36
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.bold,
+    color: darkBlueGradient,
+    fontSize: 36);
 TextStyle loginButtonStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.w700, 
-  color: lightBlue, 
-  fontSize: 24
-);
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.w700,
+    color: lightBlue,
+    fontSize: 24);
 
 TextStyle registerButtonStyle = new TextStyle(
-  fontFamily: 'Avenir', 
-  fontWeight: FontWeight.bold, 
-  color: red, 
-  fontSize: 24
-);
-
+    fontFamily: 'Segoe UI',
+    fontWeight: FontWeight.bold,
+    color: red,
+    fontSize: 24);
