@@ -1,29 +1,39 @@
+import 'package:todolist/models/subtasks.dart';
+
 class Task {
   String title; //project title
-  int taskId; //project Id
+  int id; //project Id
   int index;
   String taskKey;
-  List<Task> tasks; // tasks associated with the project
-  String group; // which group added that project
+  List<Subtask> subtasks; // tasks associated with the project
+  String groupKey; // which group added that project
+  String groupName;
   String note;
-  DateTime timeToComplete;
   bool completed; // has all tasks within the project been completed
   String repeats;
   List<DateTime> reminders;
   //DateTime deadline;
 
-  Task(this.title, this.group, this.completed, this.taskId, this.note,
-      this.taskKey, this.index);
+  Task(
+      {this.title,
+      this.groupKey,
+      this.groupName,
+      this.completed,
+      this.id,
+      this.note,
+      this.taskKey,
+      this.index});
 
   factory Task.fromJson(Map<String, dynamic> parsedJson) {
     return Task(
-      parsedJson['title'],
-      parsedJson['group'],
-      parsedJson['completed'],
-      parsedJson['id'],
-      parsedJson['note'],
-      parsedJson['task_key'],
-      parsedJson['index'],
+      id: parsedJson['id'],
+      title: parsedJson['title'],
+      index: parsedJson['index'],
+      taskKey: parsedJson['task_key'],
+      completed: parsedJson['completed'],
+      note: parsedJson['note'],
+      groupKey: parsedJson['group_key'],
+      groupName: parsedJson['group_name'],
     );
   }
 }
